+++ packages/client-connectivity/src/sdk/sdk.ts
@@ -104,6 +104,7 @@ export class ConnectivityClient extends ClientSDK {
                 path: path$,
                 headers: headers$,
                 query: query$,
+                timeoutMs: options?.timeoutMs || this.options$.timeoutMs || -1,
             },
             options
         );