+++ packages/client-connectivity/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [typescript v1.0.0-alpha.3] packages/client-connectivity
 ### Releases
 - [NPM v1.0.0-alpha.3] https://www.npmjs.com/package/@formance/sdk-connectivity/v/1.0.0-alpha.3 - packages/client-connectivity
+
+## 2024-07-12 00:21:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.332.1 (2.366.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.4] packages/client-connectivity
+### Releases
+- [NPM v1.0.0-alpha.4] https://www.npmjs.com/package/@formance/sdk-connectivity/v/1.0.0-alpha.4 - packages/client-connectivity